fix bug in uScriptAct_SetInt.. it was outputting a float instead of an int
git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@740 b191eb17-2330-428c-a13f-159e02ee8e09

+++ uScript_Dev/Assets/uScript/uScriptEditor/Nodes/Actions/Set Variable/uScriptAct_SetInt.cs
@@ -18,7 +18,7 @@ public class uScriptAct_SetInt : uScriptLogic
 {
    public bool Out { get { return true; } }
 
-   public void In(int Value, out float Target)
+   public void In(int Value, out int Target)
    {
       Target = Value;
    }
